refactor: consolidate grades search into gradesSearch module
- Single gradesSearch.ts facade: prefetchSearchGrades, useSectionGpa, lookup helpers
- GpaCell delegates to useSectionGpa hook
- Remove hydrateGrades.ts
