+++ apps/antalmanac/src/components/RightPane/SectionTable/SectionTableBody/SectionTableBodyCells/GpaCell.tsx
@@ -1,9 +1,9 @@
 import { TableBodyCellContainer } from '$components/RightPane/SectionTable/SectionTableBody/SectionTableBodyCells/TableBodyCellContainer';
 import { GradesPopover } from '$components/RightPane/SectionTable/SectionTablePopover/GradesPopover';
 import { useIsMobile } from '$hooks/useIsMobile';
-import { trpcReact } from '$lib/api/trpc';
+import { useSectionGpa } from '$lib/gradesSearch';
 import { ButtonBase, Popover, useTheme } from '@mui/material';
-import { useCallback, useMemo, useState } from 'react';
+import { useCallback, useState } from 'react';
 
 interface GpaCellProps {
     deptCode: string;
@@ -11,48 +11,11 @@ interface GpaCellProps {
     instructors: string[];
 }
 
-function formatGpa(averageGPA: number | null | undefined): string {
-    if (averageGPA == null) {
-        return '';
-    }
-    return averageGPA.toFixed(2);
-}
-
 export const GpaCell = ({ deptCode, courseNumber, instructors }: GpaCellProps) => {
     const isMobile = useIsMobile();
     const theme = useTheme();
-    const utils = trpcReact.useUtils();
     const [anchorEl, setAnchorEl] = useState<Element>();
-
-    const namedInstructors = useMemo(() => instructors.filter((i) => i !== 'STAFF'), [instructors]);
-
-    const cachedMatch = useMemo(() => {
-        for (const instructor of namedInstructors) {
-            const cached = utils.grades.aggregateGrades.getData({
-                department: deptCode,
-                courseNumber,
-                instructor,
-            });
-            if (cached?.gradeDistribution?.averageGPA != null) {
-                return { instructor, gpa: formatGpa(cached.gradeDistribution.averageGPA) };
-            }
-        }
-        return null;
-    }, [courseNumber, deptCode, namedInstructors, utils]);
-
-    const fallbackInstructor = namedInstructors[0] ?? '';
-
-    const { data: fetchedGrades, isLoading } = trpcReact.grades.aggregateGrades.useQuery(
-        { department: deptCode, courseNumber, instructor: fallbackInstructor },
-        {
-            enabled: !cachedMatch && !!fallbackInstructor,
-            select: (data) => data?.gradeDistribution ?? null,
-        }
-    );
-
-    const gpa = cachedMatch?.gpa ?? formatGpa(fetchedGrades?.averageGPA);
-    const instructor = cachedMatch?.instructor ?? fallbackInstructor;
-    const loading = !cachedMatch && isLoading;
+    const { gpa, instructor, loading } = useSectionGpa(deptCode, courseNumber, instructors);
 
     const handleClick = useCallback((event: React.MouseEvent<HTMLElement>) => {
         setAnchorEl((current) => (current ? undefined : event.currentTarget));
